pyproject.toml
[tool.poetry]
name = "proteinmpnn"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.0"
numpy = "^1.26.4"
biopython = "^1.83"
psutil = "^5.9.8"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120